requirements.txt
flask
requests
python-dotenv
openai
openai-whisper
pydub
gunicorn
whatsapp-wrapper
